[GALLERY] - Gallery template base
This is a PR to add a starting point for the new Gallery template. It establishes the basic project layout, as well the base for a hero partial. Sections included so far:

- Main gallery page (gallery.amp.html)
- Inner gallery page (gallery-inner.amp.html)
- Hero partial (partials/hero.snip.html)

As this is just to establish a basic layout and project structure, there are no mocks or staging links at this point, but they will be included in subsequent PRs.