[GALLERY] WIP - Sidebar component
As mentioned in #631, this is a WIP PR for a Sidebar that reuses an
Ampstart component - the `Navbar`.

I'm not sure if this approach is appropriate, so I was hoping to get
some insight from @camelburrito.

There are two main issues I see with it:

1. Using an `::after` pseudo element's `content` rule to show the
"Current Exhibitions" text, as its not properly screen-reader friendly
(as far as I know).
2. The only way to get images and SVG icons into the sidebar `links` is
to insert them as HTML in the `gallery.amp/json` file. This doesn't feel
right to me, but I'm curious what you think about it.

Anyway, let me know if this approach is ok, or if there's something
different we should we do.