[TRAVEL] Results filter (#725)
Adds the filter component for the Travel template results page. Link to component in staging:

https://travel-results-filter.firebaseapp.com/templates/travel/travel-results.amp

All travel-result-page components use a [src] bound amp-list element to pass attributes to filter the search results by. This filter component displays city, departure/return, max price, sorting and activity-type parameters, and allows you to alter activity-type, city, max-price, and sorting-parameters to alter the query. This component filters which results are displayed when paired with the "results" component.