Lock `first-mate` subdependency to v7.3.0
The `highlights` package depends on a version range of the `first-mate` package, which currently resolves to v7.4.0. This version of `first-mate` depends on `oniguruma` v7.2.0. This version of `oniguruma` is broken. I am locking `first-mate` at v7.3.0, which uses the previous functioning version of `oniguruma` (v7.1.0) in order for `npm i` to complete.